+++ src/main.ts
@@ -13,7 +13,14 @@ import {
   MarkdownFileInfo,
 } from "obsidian";
 
-import { resolve, relative, join, parse, posix, basename, dirname } from "path";
+import {
+  resolve,
+  relative,
+  join,
+  parse,
+  basename,
+  dirname,
+} from "path-browserify";
 import { existsSync, mkdirSync, writeFileSync, unlink } from "fs";
 
 import fixPath from "fix-path";
@@ -559,6 +566,7 @@ export default class imageAutoUploadPlugin extends Plugin {
       this.app.workspace.on(
         "editor-paste",
         (evt: ClipboardEvent, editor: Editor, markdownView: MarkdownView) => {
+          console.log("editor-paste", evt.clipboardData.files);
           const allowUpload = this.helper.getFrontmatterValue(
             "image-auto-upload",
             this.settings.uploadByClipSwitch